nested object method renamed for mongo data usage
