+++ WebApp.Logger/Extensions/ObjectExtension.cs
@@ -68,5 +68,52 @@ namespace WebApp.Logger.Extensions
 
             return obj;
         }
+
+        public static object ReadNestedObject(this object obj)
+        {
+            if (obj == null)
+                return null;
+
+            var typeName = obj.GetType().Name.ToLower();
+
+            //if (!HashTableSerializeExtentions.typecontainer.Contains(typeName))
+            //{
+            //    HashTableSerializeExtentions.typecontainer.Add(typeName);
+            //}
+
+            if (typeName.Contains("object") || typeName.Contains("dictionary"))
+            {
+                var hash = obj.ToJson().ToModel<Hashtable>();
+                var newHash = new Hashtable();
+
+                foreach (var key in hash.Keys)
+                {
+                    var newHashValue = ReadNestedObject(hash[key]);
+                    newHash.Add(key, newHashValue);
+                }
+
+                return newHash;
+            }
+            else if (typeName.Contains("list") || typeName.Contains("array"))
+            {
+                var list = obj.ToJson().ToModel<List<object>>();
+
+                if (list.Count == 0)
+                    return null;
+
+                var newlist = new List<object>();
+
+                foreach (var val in list)
+                {
+                    newlist.Add(ReadNestedObject(val));
+                }
+
+                return newlist;
+            }
+            else
+            {
+                return obj;
+            }
+        }
     }
 }
