+++ WebApp.Logger/Loggers/Providers/Mongos/MongoDocumentMapper.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Net;
 using WebApp.Common.Serialize;
+using WebApp.Logger.Extensions;
 using WebApp.Logger.Models;
 
 namespace WebApp.Logger.Loggers.Providers.Mongos
@@ -16,10 +17,10 @@ namespace WebApp.Logger.Loggers.Providers.Mongos
                 Type = model.Type,
                 TableName = model.TableName,
                 DateTime = model.DateTime,
-                PrimaryKey = model.PrimaryKey.IdentifyNestedObjects(),
-                OldValues = model.OldValues.IdentifyNestedObjects(),
-                NewValues = model.NewValues.IdentifyNestedObjects(),
-                AffectedColumns = model.AffectedColumns.IdentifyNestedObjects(),
+                PrimaryKey = model.PrimaryKey.ReadNestedObject(),
+                OldValues = model.OldValues.ReadNestedObject(),
+                NewValues = model.NewValues.ReadNestedObject(),
+                AffectedColumns = model.AffectedColumns.ReadNestedObject(),
                 CreatedBy = model.CreatedBy,
                 TraceId = model.TraceId,
                 CreatedDateUtc = model.CreatedDateUtc,
@@ -41,10 +42,10 @@ namespace WebApp.Logger.Loggers.Providers.Mongos
                 Url = model.Url,
                 Source = model.Source,
                 Form = model.Form,
-                Body = model.Body.IdentifyNestedObjects(),
-                Response = model.Response.IdentifyNestedObjects(),
-                RequestHeaders = model.RequestHeaders.IdentifyNestedObjects(),
-                ResponseHeaders = model.ResponseHeaders.IdentifyNestedObjects(),
+                Body = model.Body.ReadNestedObject(),
+                Response = model.Response.ReadNestedObject(),
+                RequestHeaders = model.RequestHeaders.ReadNestedObject(),
+                ResponseHeaders = model.ResponseHeaders.ReadNestedObject(),
                 Scheme = model.Scheme,
                 TraceId = model.TraceId,
                 Proctocol = model.Proctocol,
@@ -75,11 +76,11 @@ namespace WebApp.Logger.Loggers.Providers.Mongos
                 Area= model.Area,
                 ControllerName= model.ControllerName,
                 ClassName= model.ClassName,
-                Command= model.Command.IdentifyNestedObjects(),
+                Command= model.Command.ReadNestedObject(),
                 IpAddress= model.IpAddress,
-                Connection= model.Connection.IdentifyNestedObjects(),
+                Connection= model.Connection.ReadNestedObject(),
                 Duration= model.Duration,
-                Event= model.Event.IdentifyNestedObjects(),
+                Event= model.Event.ReadNestedObject(),
                 Host= model.Host,
                 Message= model.Message,
                 MethodName= model.MethodName,
@@ -109,10 +110,10 @@ namespace WebApp.Logger.Loggers.Providers.Mongos
                 Url = model.Url,
                 Source = model.Source,
                 Form = model.Form,
-                Body = model.Body.IdentifyNestedObjects(),
-                Response = model.Response.IdentifyNestedObjects(),
-                RequestHeaders = model.RequestHeaders.IdentifyNestedObjects(),
-                ResponseHeaders = model.ResponseHeaders.IdentifyNestedObjects(),
+                Body = model.Body.ReadNestedObject(),
+                Response = model.Response.ReadNestedObject(),
+                RequestHeaders = model.RequestHeaders.ReadNestedObject(),
+                ResponseHeaders = model.ResponseHeaders.ReadNestedObject(),
                 ErrorCode = model.ErrorCode,
                 Scheme = model.Scheme,
                 TraceId = model.TraceId,
